+++ CopilotKit/packages/react-textarea/CHANGELOG.md
@@ -1,5 +1,14 @@
 # ui
 
+## 1.8.7-next.0
+
+### Patch Changes
+
+- Updated dependencies [8b8474f]
+  - @copilotkit/runtime-client-gql@1.8.7-next.0
+  - @copilotkit/react-core@1.8.7-next.0
+  - @copilotkit/shared@1.8.7-next.0
+
 ## 1.8.6
 
 ### Patch Changes